# server/docs/swagger.yaml
definitions:
...
  github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  model.Notification:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      orderId:
        type: integer
      read:
        type: boolean
      time:
        type: string
      title:
        type: string
      type:
        description: order, system, promotion, message
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.9.0
paths:
...
  /playmate/notifications:
    get:
      consumes:
      - application/json
      parameters:
      - description: 类型
        in: query
        name: type
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 结束时间
        in: query
        name: endTime
        type: string
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Notification'
                  type: array
                pagination:
                  additionalProperties:
                    type: integer
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取通知列表
      tags:
      - Notification
  /playmate/notifications/{id}/read:
    put:
      consumes:
      - application/json
      parameters:
      - description: 通知ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 标记成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                message:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 标记通知为已读
      tags:
      - Notification
  /playmate/notifications/read-all:
    put:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 全部标记为已读
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                message:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 标记所有通知为已读
      tags:
      - Notification
  /playmate/notifications/unread-count:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_flipped-aurora_gin-vue-admin_server_plugin_playmate_model_response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: integer
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取未读通知数量
      tags:
      - Notification
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...